macOS: Don't send posted events via runloop observer
The logic was added in b0016ea9a6b225757e3ee06b50e8f7d05463ddf7,
but the test-case in the bug report works fine without this logic.
As sending events from a runloop observer is dubious, and causes
problems since it's not tied to the quite complex machinery of
the general event processing, it's better to remove it.

The original issue is already tested by sendEventsOnProcessEvents
in tst_QCore/GuiApplication.

Task-number: QTBUG-4521
Change-Id: I857519a09a13a72b7434727c0229606403be7ed3
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>